# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JOB API",
        "description": "API da plataforma JOB para intermediação de serviços.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/publicacoes/{publicacao}/anexos/{anexo}": {
            "get": {
                "tags": [
                    "Publicação - Anexos"
                ],
                "summary": "Visualiza um anexo",
                "description": "Retorna o arquivo diretamente para o navegador. Imagens, PDFs e vídeos podem ser visualizados de acordo com o suporte do navegador.",
                "operationId": "4e3bf3484318e453ed0f107c49ada39a",
                "parameters": [
                    {
                        "name": "publicacao",
                        "in": "path",
                        "description": "ID da publicação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "anexo",
                        "in": "path",
                        "description": "ID do anexo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo retornado com sucesso",
                        "content": {
                            "image/jpeg": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "image/png": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "image/webp": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "video/mp4": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "video/webm": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Publicação, anexo ou arquivo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Publicação - Anexos"
                ],
                "summary": "Remove um anexo da publicação",
                "description": "Remove o anexo do banco de dados e também o arquivo físico do armazenamento privado.",
                "operationId": "9a00b6d0322e8ef8058a1272885a25ad",
                "parameters": [
                    {
                        "name": "publicacao",
                        "in": "path",
                        "description": "ID da publicação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "anexo",
                        "in": "path",
                        "description": "ID do anexo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anexo removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Anexo removido com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Usuário não possui permissão para remover o anexo"
                    },
                    "404": {
                        "description": "Publicação ou anexo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação do Laravel Sanctum",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}